chore(set): move to `set` directory.,
