+++ src/lib/set/index.ts
@@ -1,0 +1,2 @@
+export { DataSet } from './data-set.class';
+export { WeakDataSet } from './weak-data-set.class';
